misc : changed WIN32 typedefs and functions into std's. / forward decl

+++ GraphicsPlayground/GraphicsPlayground/Engine/Source/Runtime/Engine/Classes/Engine/Asset/SkeletalMeshAsset.h
@@ -34,7 +34,7 @@ struct FSkeletalMeshRenderData
     FString DisplayName;
 
     TArray<FSkeletalMeshVertex> Vertices;
-    TArray<UINT> Indices;
+    TArray<uint16> Indices;
 
     TArray<FMaterialInfo> Materials;
     TArray<FMaterialSubset> MaterialSubsets;
